checkbox ts化
