+++ package.json
@@ -12,9 +12,11 @@
     "next": "12.1.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "sass": "^1.49.9"
+    "sass": "^1.49.9",
+    "typescript": "^4.6.2"
   },
   "devDependencies": {
+    "@types/react": "^17.0.41",
     "eslint": "8.11.0",
     "eslint-config-next": "12.1.0",
     "prettier": "^2.6.0"
