+++ types/resasApiType.ts
@@ -1,0 +1,1 @@
+export type ResasApi = {}
